Only authors can delete

+++ trainmate-back/src/main/java/com/trainmateback/trainmateback/controller/ForumController.java
@@ -2,8 +2,11 @@ package com.trainmateback.trainmateback.controller;
 
 import com.trainmateback.trainmateback.model.Answer;
 import com.trainmateback.trainmateback.model.Question;
+import com.trainmateback.trainmateback.model.TrainMateUser;
 import com.trainmateback.trainmateback.repository.QuestionController;
+import com.trainmateback.trainmateback.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -16,10 +19,13 @@ public class ForumController {
 
     @Autowired
     private QuestionController questionController;
+    @Autowired
+    private UserRepository userRepository;
 
     @PostMapping("/addQuestion")
     ResponseEntity<?> addQuestion(@RequestBody Question question){
-        Question question_aux = new Question(question.getQuestion());
+        TrainMateUser user = userRepository.findByToken(question.getToken());
+        Question question_aux = new Question(question.getQuestion(), user.getUsername());
         questionController.save(question_aux);
         return ResponseEntity.ok("");
     }
@@ -38,11 +44,16 @@ public class ForumController {
         return ResponseEntity.ok("");
     }
 
-    @DeleteMapping("deleteQuestion/{id}")
-    ResponseEntity<?> deleteQuestion(@PathVariable long id){
+    @DeleteMapping("deleteQuestion/{id}/{token}")
+    ResponseEntity<?> deleteQuestion(@PathVariable long id, @PathVariable String token){
             Question question = questionController.findById(id);
-            questionController.delete(question);
-            return ResponseEntity.ok("");
-        }
+            TrainMateUser user = userRepository.findByToken(token);
+            if (user.getUsername() == question.getAuthor()) {
+                questionController.delete(question);
+                return ResponseEntity.ok("");
+            }
+            else return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("");
+
+    }
 
 }
